chore: 公共函数提取，移除无用代码
