+++ web/main/index.ts
@@ -1,10 +1,7 @@
 import '../app.scss';
-// import '../app.css';
 import '@ggchivalrous/db-ui/components/theme/index.css';
 
 import App from './app.svelte';
-
-// storePopup.set({ computePosition, autoUpdate, offset, shift, flip, arrow });
 
 const app = new App({
   target: document.getElementById('app'),
