+++ web/util/util.ts
@@ -71,3 +71,40 @@ export const loadImage = async (url: string) => {
 
   return promise1;
 };
+
+// 计算图片整体亮度
+export const calcAverageBrightness = (ctx: CanvasRenderingContext2D, width: number, height: number) => {
+  const imageData = ctx.getImageData(0, 0, width, height);
+  const { data } = imageData;
+  let totalBrightness = 0;
+
+  for (let i = 0; i < data.length; i += 4) {
+  // 简单的亮度计算方法
+    const brightness = (data[i] + data[i + 1] + data[i + 2]) / 3;
+    totalBrightness += brightness;
+  }
+  return totalBrightness / (width * height);
+};
+
+export interface IFontInfo {
+  name: string
+  path: string
+}
+
+// 导入字体
+export async function importFont(arr: IFontInfo[]) {
+  for (const i of arr) {
+    const font = new FontFace(i.name, `url('${i.path}')`);
+    const _font = await font.load().catch((e) => console.log('%s 字体加载失败', i.name, e));
+    if (_font) {
+      (document.fonts as any).add(_font);
+    }
+  }
+}
+
+export const createCanvas = (w: number, h: number) => {
+  const _canvas = document.createElement('canvas');
+  _canvas.width = w;
+  _canvas.height = h;
+  return _canvas;
+};
